Added header for DoLahMain
also moved Command header to the commands folder
