+++ DoLah/DoLah.vcxproj
@@ -125,6 +125,8 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="basic.h" />
+    <ClInclude Include="Command.h" />
+    <ClInclude Include="DoLah.h" />
     <ClInclude Include="Models\Task.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
